+++ src/GlassCat/Air.jl
@@ -11,3 +11,8 @@ function info(io::IO, ::AirType)
     println(io, "GlassCat.Air")
     println(io, "Material representing air, RI is always 1.0 at system temperature and pressure, absorption is always 0.0.")
 end
+
+"""
+Special glass to represent air. Refractive index is defined to always be 1.0 for any temperature and pressure (other indices are relative to this).
+"""
+const Air = AirType()
